contentunderstandingdemo/requirements.txt
azure-identity
python-dotenv
requests
azure-ai-projects